Repository onion and code cleanup
