+++ src/Chirp.Core/Chirp.Core.csproj
@@ -8,8 +8,6 @@
 
   <ItemGroup>
     <Compile Remove="Classes\CheepRepository.cs" />
-    <Compile Remove="Helpers\ChirpDBContext.cs" />
-    <Compile Remove="Interfaces\ICheepRepository.cs" />
   </ItemGroup>
 
   <ItemGroup>
@@ -26,9 +24,5 @@
         <PrivateAssets>all</PrivateAssets>
       </PackageReference>
       <PackageReference Include="Microsoft.EntityFrameworkCore.SQLite" Version="8.0.8" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="Helpers\" />
     </ItemGroup>
 </Project>
