+++ src/Chirp.Razor/Pages/Shared/_CheepListPartial.cshtml
@@ -1,8 +1,9 @@
-@using Chirp.Core.Classes;
-@using Chirp.Core.Interfaces;
-@using System.Text;
-@using System.Security.Cryptography;
-@using Chirp.Services.Interfaces;
+@	using Chirp.Core.Classes;
+@	using Chirp.Core.Interfaces;
+@	using System.Text;
+@	using System.Security.Cryptography;
+@	using Chirp.Services.Interfaces;
+	using Chirp.Core.DTO;
 @model (List<CheepDTO> Cheeps, Range CheepRange, int PageNumber, int TotalPages, Author? UserAuthor, List<Author> FollowedAuthors)
 @inject ICheepService CheepService
 @{
