+++ test/Chirp.Razor.Tests/Tests/UnitTests.cs
@@ -1,4 +1,5 @@
 ﻿using Chirp.Core.Classes;
+using Chirp.Core.DTO;
 using Chirp.Repositories.Repositories;
 using Chirp.Services;
 using Chirp.Tests.Helpers;
